+++ app/models/assignment.rb
@@ -160,6 +160,14 @@ class Assignment < Assessment
     SectionDueDate.due_date_for(section, self)
   end
 
+  # Return the start_time for +section+ if it is not nil, otherwise return this
+  # assignments start_time instead.
+  def section_start_time(section)
+    return start_time unless section_due_dates_type
+
+    section&.section_due_dates&.find_by(assignment: self)&.start_time || start_time
+  end
+
   # Calculate the latest due date among all sections for the assignment.
   def latest_due_date
     return due_date unless section_due_dates_type
@@ -193,7 +201,7 @@ class Assignment < Assessment
 
     due_dates = Hash.new { |h, k| h[k] = due_date }
     section_due_dates.each do |grouping_id, sec_due_date|
-      due_dates[grouping_id] = sec_due_date
+      due_dates[grouping_id] = sec_due_date unless sec_due_date.nil?
     end
     grouping_extensions.each do |grouping_id, ext|
       due_dates[grouping_id] += ActiveSupport::Duration.parse(ext)
@@ -1005,10 +1013,12 @@ class Assignment < Assessment
   # (Basically, it's nice for a group to share a repo among assignments, but at a certain point during the course
   # we may want to add or [more frequently] remove some students from it)
   def self.get_repo_auth_records
-    Assignment.joins(:assignment_properties)
-              .includes(groupings: [:group, { accepted_student_memberships: :user }])
-              .where(assignment_properties: { vcs_submit: true })
-              .order(due_date: :desc)
+    records = Assignment.joins(:assignment_properties)
+                        .includes(groupings: [:group, { accepted_student_memberships: :user }])
+                        .where(assignment_properties: { vcs_submit: true })
+                        .order(due_date: :desc)
+    records.where(assignment_properties: { is_timed: false })
+           .or(records.where.not(groupings: { start_time: nil }))
   end
 
   ### /REPO ###