+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_05_26_185329) do
+ActiveRecord::Schema.define(version: 2020_06_08_190551) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -123,6 +123,9 @@ ActiveRecord::Schema.define(version: 2020_05_26_185329) do
     t.boolean "hide_unassigned_criteria", default: false, null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.string "duration"
+    t.datetime "start_time"
+    t.boolean "is_timed", default: false, null: false
     t.index ["assessment_id"], name: "index_assignment_properties_on_assessment_id", unique: true
   end
 
@@ -279,6 +282,7 @@ ActiveRecord::Schema.define(version: 2020_05_26_185329) do
     t.integer "test_tokens", default: 0, null: false
     t.text "starter_code_revision_identifier"
     t.bigint "assessment_id", null: false
+    t.datetime "start_time"
     t.index ["assessment_id", "group_id"], name: "groupings_u1", unique: true
   end
 
@@ -419,6 +423,7 @@ ActiveRecord::Schema.define(version: 2020_05_26_185329) do
     t.datetime "due_date"
     t.integer "section_id"
     t.bigint "assessment_id"
+    t.datetime "start_time"
   end
 
   create_table "sections", id: :serial, force: :cascade do |t|