+++ Changelog.md
@@ -18,6 +18,7 @@
 - Fixed bug where image annotations did not stay fixed relative to the image (#4706)
 - Fixed bug where image annotations did not load properly (#4706)
 - Fixed bug where downloading files in nested directories renamed the downloaded file (#4730)
+- Introduces timed assignments (#4665)
 
 ## [v1.9.3]
 - Fixed inverse association bug with assignments (#4551)